LUS-554 lprocfs: Increase OBD_HIST_MAX to 256
Increase OBD_HIST_MAX to 256 to match OSC_MAX_RIF_MAX so
that the rpcs in flight histogram which is displayed by
reading the rpc_stats proc interface is correct.

Note the memory cost of this is:
sizeof(unsigned long)*(OBD_HIST_MAX_new -
OBD_HIST_MAX_old)*OST_count

For a 64 bit system with 2048 OSTs, this is:

8 bytes*256-32*2048 = 3.5 MiB

Which is trivial.

Signed-off-by: Chris Horn <hornc@cray.com>
Change-Id: Iabbea849960d5c9a67c52fdc32bd7a3d6d041426
Reviewed-on: http://es-gerrit.dev.cray.com:8080/17273
Reviewed-by: Alexander Boyko <c17825@cray.com>
Reviewed-by: Alexander Zarochentsev <c17826@cray.com>
Tested-by: Elena Gryaznova <c17455@cray.com>