MRP-2723 test: Testing Round-Robin allocation
Adding a test for fix in MRP-2471(LU-977) which shows that
in absence of protection to lqr_start_idx there is
possibility of imblance in allocating objects on OSTs
with round-robin algorithm.

This test checks the protection of lqr_start_idx by using
a new reproducer, rr_alloc which uses MPI to create files in
parallel, and checking the even file distribution over OSTs.

Distribution check formula is adjusted as per implementation,i.e
some factors like - exhaustion of pre-created objects and counter
‘lrq_start_count’(lod_qos.c) reseed are also taken care so that
object allocation is not affected by these factors.

Signed-off-by: Ashish Maurya <ashish.maurya@seagate.com>
Signed-off-by: Rahul Deshmukh <rahul.deshmukh@seagate.com>
Change-Id: I2b43fd928585dabe6c356fcaa76f944bb8c7877d
Reviewed-on: http://es-gerrit.xyus.xyratex.com:8080/10776
Reviewed-by: Vladimir Saveliev <vladimir.saveliev@seagate.com>
Reviewed-by: Vikram Jadhav <jadhav.vikram@seagate.com>
Tested-by: Elena V. Gryaznova <elena.gryaznova@seagate.com>