MRP-3666 hsm: issue fixes/improvement related to MRP-2464
One of the calls to hsm_cancel_all_actions() comes from
mdt_hsm_agent_unregister() during kill event of a copytool.
mdt_hsm_agent_unregister() 1st removes the copy tool (hsm_agent)
from the list and then makes call to hsm_cancel_all_actions().

So when the call lands into hsm_cancel_all_actions(),
it is sure that the copy tool is no more and mdt_hsm_agent_send()
returns error = -2 as it tries to find the copy tool to send
the hsm cancel request but we already removed it.

In this case, if the copy tool agent is already unregistered then
we are just going to add the hsm cancel request to agent record
and mark request state as ARS_CANCELED. mdt_cordinator()
thread running in the back ground looks into the request state
and marks it as CANCELED.

Signed-off-by: vinayakswami hariharmath <vinayakswami.hariharmath@seagate.com>
Change-Id: Ib8c1b014c6d3b309fc60ce8707499cbfe99506ab
Reviewed-on: http://es-gerrit.xyus.xyratex.com:8080/13279
Reviewed-by: Artem Blagodarenko <artem.blagodarenko@seagate.com>
Reviewed-by: Sergey Cheremencev <sergey.cheremencev@seagate.com>
Reviewed-by: Parinay Vijayprakash Kondekar <parinay.kondekar@seagate.com>
Tested-by: Parinay Vijayprakash Kondekar <parinay.kondekar@seagate.com>