MRP-3293 mdd: move lfsck_stop after stack cleanup
The scenario:
lfsck_stop waits for the lfsck master engine process to complete.
The lfsck master engine process in turn waits for the FLD_QUERY
request RPC to complete.
The FLD_QUERY RPC is waiting for a reply from the remote server(MDT0)
and the remote server has already been unmounted/is offline. Hence,
this RPC will be blocked until the remote server is online.
Due to the RPC being blocked, lfsck_stop waits infinitely for the
lfsck to stop on the target.

The fix:
lfsck_stop should occur after all dependent parts of the stack
finish cleanup.This will avoid the blocking of fld requests.

Signed-off-by: Noopur Maheshwari <noopur.maheshwari@seagate.com>
Signed-off-by: Alexey Leonidovich Lyashkov <alexey.lyashkov@seagate.com>
Change-Id: I72588d25525b00a353fb7c0e0a65e0beb3c2b327
Reviewed-on: http://es-gerrit.xyus.xyratex.com:8080/10840
Reviewed-by: Rahul Deshmukh <rahul.deshmukh@seagate.com>
Tested-by: Elena V. Gryaznova <elena.gryaznova@seagate.com>