MRP-2928 test: allow time for copytools to stop
In sanity-hsm allow up to 200 seconds for the copytool to stop. This
is needed to prevent sporadic failures due to slow NFS (for the
copytool log file) delaying the termination of the copytool.

Also as pkill results in calling mdt_hsm_ct_unregister() which invokes
hsm_cancel_all_actions() (because of the patch MRP-2464) which does
clean up of pending hsm requests, so pkill takes some time to finish.

Intel-bug-id: LU-7136
Signed-off-by: John L. Hammond <john.hammond@intel.com>
Signed-off-by: vinayakswami hariharmath <vinayakswami.hariharmath@seagate.com>
Change-Id: I0b99d8c4ad4040abaed0a5ded15b51dc13fa4daf
Reviewed-on: http://es-gerrit.xyus.xyratex.com:8080/10845
Reviewed-by: Mirza Arshad Mirza Hussain <arshad.hussain@seagate.com>
Reviewed-by: Ashish Purkar <ashish.purkar@seagate.com>
Tested-by: Parinay Vijayprakash Kondekar <parinay.kondekar@seagate.com>