MRP-3936 llite: Stop file creation for ro bind mnt
While remounting the bind mount of lustre with read-only, vfs
sets MNT_READONLY in the mnt flags and does not make a call
to lustre. Hence, the change in mnt flags is not reflected in
lustre.

Therefore, file creation goes ahead in lookup operation of
lustre with LOOKUP_CREATE intent set and converted to IT_CREAT.

The fix is to disallow file creation by not setting IT_CREAT
intent when bind mnt pt is readonly and unsetting O_CREAT.

Signed-off-by: Noopur Maheshwari <noopur.maheshwari@seagate.com>
Change-Id: Ia96d6fa9d310942c96de01ed8bac88cf170cec57
Reviewed-on: http://es-gerrit.xyus.xyratex.com:8080/12987
Reviewed-by: Ashish Purkar <ashish.purkar@seagate.com>
Reviewed-by: Alexander Zarochentsev <alexander.zarochentsev@seagate.com>
Tested-by: Elena V. Gryaznova <elena.gryaznova@seagate.com>