MRP-4138 ldiskfs: enable quota journalling
This commit enables proper quota journalling
if quotas are activated using mkfs -O quota,
which is Lustre default.

ldiskfs_mark_dquot_dirty() should not call
ldiskfs_write_dquot() if the dquot is already
dirty since we are guaranteed that another
thread is already writing the dquot and
will update the buffer with the latest
dquot data.

This commit is an improved version of
LU-8216 by Wang Shilong <wshilong@ddn.com>.

The original version of the patch led to
nearly 15-20% performance drop in mdtest and
had to be reworked.

Change-Id: I969b86d34e02f2bd731fb62692f1dd1e094a957f
Signed-off-by: Andrew Perepechko <andrew.perepechko@seagate.com>
Reviewed-on: http://es-gerrit.xyus.xyratex.com:8080/14620
Reviewed-by: Alexander Zarochentsev <alexander.zarochentsev@seagate.com>
Reviewed-by: Alexey Leonidovich Lyashkov <alexey.lyashkov@seagate.com>
Tested-by: Elena V. Gryaznova <elena.gryaznova@seagate.com>