MRP-4267 ldlm: glimpse each client once
At scale, we need the fix to glimpse each client only once,
or we can end up sending too many glimpse requests.

The easiest way to pull this in was to take the upstream
version of ofd_intent_cb exactly.  This restructures the
code a bit, but the only semantic difference is the
addition of the code block which does this check:
gl_work->gl_lock->l_export == victim_lock->l_export

Also cleaned up discarding the ldlm_glimpse_locks rc,
also to match upstream.

Signed-off-by: Patrick Farrell <paf@cray.com>
Change-Id: I561a654921ae9e47ba30f44da18032e33f801781
Reviewed-on: http://es-gerrit.xyus.xyratex.com:8080/14792
Reviewed-by: Vitaly Fertman <vitaly.fertman@seagate.com>
Reviewed-by: Alexey Leonidovich Lyashkov <alexey.lyashkov@seagate.com>
Tested-by: Vitaly Fertman <vitaly.fertman@seagate.com>