MRP-4191 tests: sanity/313 fail_loc is different in 2.x
the patch for MRP-3609 when upstreamed, resulted in change of
FAIL_LOC for OBD_FAIL_TGT_RCVD_EIO. Also the test case number
on 2.x was different compared to the one on master.

The patch takes care of these discrepancies, which are causing
test to fail in case of interop.

The patch also includes the changes done upstream, and are missing
on 2.x.

Test-Parameters: testlist=sanity/313
Change-Id: I00019253aa58b6bb9396256e254818adcf0b6b4b
Signed-off-by: Parinay Kondekar <Parinay.Kondekar@seagate.com>
Reviewed-on: http://es-gerrit.xyus.xyratex.com:8080/14411/
Reviewed-by: Artem Blagodarenko <artem.blagodarenko@seagate.com>
Reviewed-by: Ashish Purkar <ashish.purkar@seagate.com>
Tested-by: Elena V. Gryaznova <elena.gryaznova@seagate.com>