MRP-3752 tests: skip non-crucial LFSCK intermediateness check
Sometimes, because of system schedule issue, the test scripts
may miss to check some LFSCK intermediate status, but it will
not affect the LFSCK to auto restart after MDS remount.
We can skip some non-crucial status check, but check the LFSCK
auto restarted correctly after MDS remount.

The fix is on the lines of LU-8810/SEA-206.

Intel-bug-id: LU-8810
Signed-off-by: Noopur Maheshwari <noopur.maheshwari@seagate.com>
Signed-off-by: Fan Yong <fan.yong@intel.com>
Change-Id: I2c15adc476c1c848a35a3ece16f75b6ed72d6f2b
Reviewed-on: http://es-gerrit.xyus.xyratex.com:8080/15260
Reviewed-by: Parinay Vijayprakash Kondekar <parinay.kondekar@seagate.com>
Reviewed-by: Alexander Zarochentsev <alexander.zarochentsev@seagate.com>
Tested-by: Elena V. Gryaznova <elena.gryaznova@seagate.com>