MRP-4382 tests: set mgsfailover_dev for combined mgs/mds.
Issue:
When facet=mgsfailover, the value of mgsfailover_dev is undefined.
This leads to an error while mounting the facet.

Cause:
Since ${facet}failover_dev is set in cfg file, in a combined mgs/mds
setup, we do not define MGSDEV in the cfg file and the yaml cmd.
Hence, mgsfailover_dev remains undefined.

Fix:
Assign mgsfailover_dev with mds1failover_dev when mgs/mds are combined
while mounting facets, as mds1failover_dev is already defined when
mgs/mds are combined.

Signed-off-by: Noopur Maheshwari <noopur.maheshwari@seagate.com>
Change-Id: I2b60c55704158487136014c1d8918c1db0330ac0
Reviewed-on: http://es-gerrit.xyus.xyratex.com:8080/16043
Reviewed-by: Mirza Arshad Hussain <arshad.hussain@seagate.com>
Reviewed-by: Elena V. Gryaznova <elena.gryaznova@seagate.com>
Tested-by: Elena V. Gryaznova <elena.gryaznova@seagate.com>