LU-6952 mount: Mount options parsing problem
The code which parse the mount option expect that the default
mount options come first and then the user given options. But
actually the default options are appended after the user given
option. Because of this user given options are cleared.
Following code fix this bug.

Added a conf-sanity test 104 to verify this fix.

conf-sanity 104 was added as MRP-4162 which is upstreamed as
conf-sanity 107, moving the existing test to 107.

Test-Parameters: layout=quartet_dne_singlemds
Intel-bug-id: MRP-2193
Signed-off-by: Chennaiah Palla <chennaiah.palla@seagate.com>
Signed-off-by: Gryaznova Elena <c17455@cray.com>
Change-Id: Ia5577886b52fce0e099871abd73adbba006d0df4
Reviewed-on: http://es-gerrit.dev.cray.com:8080/16588
Reviewed-by: Vitaly Fertman <c17818@cray.com>