LUS-5578 tests: data corruption due to RPC reordering
This commit adds a new test case. The client
can receive a reply from from the original RPC
or the resent one, clearing PG_Writeback and
making it possible to modify the same page on
the client. However, the resent RPC or the
original RPC can proceed after this modification.
Eventually, stale data gets to disk.

Change-Id: Id9880569f19fc02390f5b849d25fa69a47efa51b
Signed-off-by: Andrew Perepechko <c17827@cray.com>
Test-Parameters: testlist=recovery-small/135
Reviewed-on: https://es-gerrit.dev.cray.com/153539
Reviewed-by: Alexander Zarochentsev <c17826@cray.com>
Reviewed-by: Elena Gryaznova <c17455@cray.com>
Tested-by: Elena Gryaznova <c17455@cray.com>