LUS-5578 tests: data corruption due to RPC reordering
This commit adds a new test case. The client
can receive a reply from from the original RPC
or the resent one, clearing PG_Writeback and
making it possible to modify the same page on
the client. However, the resent RPC or the
original RPC can proceed after this modification.
Eventually, stale data gets to disk.

Change-Id: I47314bcfc005aa513fd0d5202f3a5eb5090e181a
Signed-off-by: Andrew Perepechko <c17827@cray.com>
Test-Parameters: testlist=recovery-small/135
Reviewed-on: https://es-gerrit.dev.cray.com/153669
Reviewed-by: Alexander Zarochentsev <c17826@cray.com>
Reviewed-by: Alexey Lyashkov <c17817@cray.com>
Tested-by: Elena Gryaznova <c17455@cray.com>