LUS-5578 ofd: data corruption due to RPC reordering
Without read-only cache, it is possible that a client
resends a BRW RPC, receives a reply from the original
BRW RPC, modifies the same data and sends a new BRW
RPC, however, because of RPC reordering stale data
gets to disk.

Let's use range locking to protect against this race.

Change-Id: Iea8f5c52c650c1112c97935968da945c96c8bd54
Signed-off-by: Andrew Perepechko <c17827@cray.com>
Reviewed-on: https://es-gerrit.dev.cray.com/153670
Reviewed-by: Alexander Zarochentsev <c17826@cray.com>
Reviewed-by: Alexey Lyashkov <c17817@cray.com>
Tested-by: Elena Gryaznova <c17455@cray.com>