LU-11195 lod: Mark comps cached on replay of layout change
Replay of a layout change request on a PFL file leaves the object
in an unexpected state: Some components can have llc_stripe set
but ldo_comp_cached is not set in the object. The next layout
change request on the same object will LBUG when it tries to free
the comp entries.

The fix is to set ldo_comp_cached on replay so subsequent layout
change requests will use the in memory components rather than
fetching them from disk.

Cray-bug-id: LUS-6227
Signed-off-by: Ann Koehler <amk@cray.com>
Change-Id: I4393910bddd08d870d08ecc087a398efcba1bbfc
Reviewed-on: https://es-gerrit.dev.cray.com/153732
Reviewed-by: Patrick Farrell <paf@cray.com>
Reviewed-by: Andrew Perepechko <c17827@cray.com>
Tested-by: Elena Gryaznova <c17455@cray.com>