LU-11473 doc: add lfs-getsom man page
The Lazy Size on MDT feature added a flag to 'lfs' to allow
users to get the LSOM data. Add the lfs "getsom" flag to the
lfs man page.

Also, make a minor correction to the llsom_sync man page that
corrects the order of the --device flag for changelog_register
and changelog_deregister.

Test-Parameters: trivial
Signed-off-by: James Nunez <jnunez@whamcloud.com>
Change-Id: I9a5acab6de3b7f32eb246c94e9975e30f63f10e1
Reviewed-on: https://review.whamcloud.com/33290
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Yingjin Qian <qian@ddn.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>