LUS-6460 llite: Lock inode on tiny write if setuid/setgid set
During a write, the setuid/setgid bits must be reset if they are
enabled and the user does not have the correct permissions. Setting
any file attributes, including setuid and setgid, requires the inode
to be locked. Writes became lockless with the introduction of
LU-1669. Locking the inode in the setuid/setgid case was added to
vvp_io_write_start() as a special case. The inode locking was not
included when support for tiny writes was added with LU-9409. This
mod adds the necessary inode lock/unlock calls to ll_do_tiny_write().

If the inode is not locked when setuid/setgid are reset, the kernel
will issue a one time warning and Lustre may hang trying to get the
inode lock in ll_setattr_raw().

This is upstream bug LU-11944.

Signed-off-by: Ann Koehler <amk@cray.com>
Change-Id: Ifefa860d0d89a500c12fcb76f612df603e03017e
Reviewed-on: https://es-gerrit.dev.cray.com/154743
Reviewed-by: Vladimir Saveliev <c17830@cray.com>
Reviewed-by: Ben Evans <jevans@cray.com>
Tested-by: Elena Gryaznova <c17455@cray.com>