LUS-7180 lod: check start index if pool is retained
When lfs setstipe -i applies to a directory which includes a pool in
layout, new start index should be checked against the pool. Otherwise,
subsequent creation will fail.

Add the check for start offset against retained pool. Have
lod_xattr_set() to return EINVAL if the start offset is not in the
pool.

Signed-off-by: Vladimir Saveliev <c17830@cray.com>
Change-Id: I84495cc3abf0ca23f418289936eca49d6f2267cf
Reviewed-on: https://es-gerrit.dev.cray.com/154899
Reviewed-by: Alexander Boyko <c17825@cray.com>
Reviewed-by: Alexey Lyashkov <c17817@cray.com>
Tested-by: Elena Gryaznova <c17455@cray.com>